// front/utils/types/form.ts
export type SizeType = {
  id?: string;
  size: string;
  is?: boolean;
};

export type ColorType = {
  id: string;
  color: string;
  colorName: string;
};

export type StyleType = {
  id: string;
  style: string;
  styleName: string;
};

export type MaterialType = {
  id: string;
  material: string;
  materialName: string;
};